Moved room service countdown logic from UI to core C++ gameplay system
Key Changes

Removed temporary UI-based timer implementation.

Implemented actual time decrement in the room service C++ object.

Synced UI with authoritative gameplay state.

+++ Source/Don/Public/Data/RoomServiceAsset.h
@@ -26,7 +26,7 @@ struct FRoomServiceOrder
 	int32 ServiceLevel = 1;
 	
 	UPROPERTY(BlueprintReadOnly, EditAnywhere)
-	int32 LimitTime = 10;
+	float LimitTime = 10;
 
 	UPROPERTY(BlueprintReadOnly, EditAnywhere)
 	int32 RewardMoney;
